backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
minio==7.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.1.0
pyyaml==6.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1